improved pd calculation

+++ src/eval/results/detection/joined.cpp
@@ -58,7 +58,7 @@ namespace EvaluationRequirementResult
         if (sum_uis_)
         {
             assert (sum_uis_ > max_gap_uis_ + no_ref_uis_);
-            pd_ = 1.0 - (missed_uis_/(sum_uis_ - max_gap_uis_ - no_ref_uis_));
+            pd_ = 1.0 - ((float)missed_uis_/(float)(sum_uis_ - max_gap_uis_ - no_ref_uis_));
             has_pd_ = true;
         }
         else
