+++ src/eval/results/detection/joined.h
@@ -31,10 +31,10 @@ namespace EvaluationRequirementResult
                 const EvaluationResultsReport::SectionContentTable& table, const QVariant& annotation) override;
 
     protected:
-        float sum_uis_ {0};
-        float missed_uis_ {0};
-        float max_gap_uis_ {0};
-        float no_ref_uis_ {0};
+        int sum_uis_ {0};
+        int missed_uis_ {0};
+        int max_gap_uis_ {0};
+        int no_ref_uis_ {0};
 
         bool has_pd_ {false};
         float pd_{0};
